# Speed up clutter and improve run.sh