This commit changes a few things:

* Use a newer commit of clutter for some extra optimizations
* Use the newer commit of text for the toLower optimization
* Run `stack exec hyperfine ...` instead of `hyperfine "stack exec ..."`, so this now no longer counts the time that stack takes to start your program
* Add an option for when nix is not installed, but hyperfine is (like on my machine)

Results on my machine:

```
Benchmark 1: "BufwiseClutter" <../kjvbible_x10.txt
  Time (mean ± σ):     584.3 ms ±   5.2 ms    [User: 553.9 ms, System: 21.7 ms]
  Range (min … max):   573.3 ms … 590.9 ms    10 runs
 
Benchmark 1: "BufwiseVHBS" <../kjvbible_x10.txt
  Time (mean ± σ):      1.007 s ±  0.010 s    [User: 0.979 s, System: 0.021 s]
  Range (min … max):    0.993 s …  1.023 s    10 runs
 
Benchmark 1: "LazyHMBS" <../kjvbible_x10.txt
  Time (mean ± σ):      4.138 s ±  0.040 s    [User: 3.742 s, System: 0.389 s]
  Range (min … max):    4.097 s …  4.235 s    10 runs
 
Benchmark 1: "LazyHMRef" <../kjvbible_x10.txt
  Time (mean ± σ):      2.323 s ±  0.017 s    [User: 2.237 s, System: 0.079 s]
  Range (min … max):    2.300 s …  2.350 s    10 runs
 
Benchmark 1: "LazyMap" <../kjvbible_x10.txt
  Time (mean ± σ):      4.763 s ±  0.029 s    [User: 4.669 s, System: 0.085 s]
  Range (min … max):    4.735 s …  4.840 s    10 runs
 
Benchmark 1: "LazyVH" <../kjvbible_x10.txt
  Time (mean ± σ):      1.969 s ±  0.013 s    [User: 1.894 s, System: 0.069 s]
  Range (min … max):    1.951 s …  1.986 s    10 runs
 
Benchmark 1: "LinewiseClutter" <../kjvbible_x10.txt
  Time (mean ± σ):     760.7 ms ±   9.1 ms    [User: 727.0 ms, System: 23.9 ms]
  Range (min … max):   746.2 ms … 780.7 ms    10 runs
 
Benchmark 1: "LinewiseVH" <../kjvbible_x10.txt
  Time (mean ± σ):      2.014 s ±  0.018 s    [User: 1.975 s, System: 0.029 s]
  Range (min … max):    1.991 s …  2.045 s    10 runs
 
Benchmark 1: "StreamingHMBS" <../kjvbible_x10.txt
  Time (mean ± σ):      4.714 s ±  0.093 s    [User: 4.666 s, System: 0.040 s]
  Range (min … max):    4.625 s …  4.883 s    10 runs
 
Benchmark 1: "StreamlyHMRef" <../kjvbible_x10.txt
  Time (mean ± σ):      2.096 s ±  0.053 s    [User: 2.065 s, System: 0.022 s]
  Range (min … max):    2.022 s …  2.191 s    10 runs
 
Benchmark 1: "StreamlyThreadedHMRef" <../kjvbible_x10.txt
  Time (mean ± σ):      3.115 s ±  0.019 s    [User: 8.371 s, System: 1.583 s]
  Range (min … max):    3.079 s …  3.145 s    10 runs
 
Benchmark 1: "StreamlyVH" <../kjvbible_x10.txt
  Time (mean ± σ):      1.875 s ±  0.014 s    [User: 1.846 s, System: 0.021 s]
  Range (min … max):    1.854 s …  1.905 s    10 runs
```