@unhammer 2022-08-08
Thanks! It got even faster =D

You're right, I wasn't watching, good of you to ping me (always thought ones own fork was automatically watched but I guess not)